# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management API",
        "description": "API documentation for the project management platform",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/github/redirect": {
            "get": {
                "tags": [
                    "Auth - GitHub"
                ],
                "summary": "Redirect to the GitHub authorization page",
                "description": "Redirects the user to the GitHub OAuth page to authorize the application.",
                "operationId": "142ff0f591d1698e9f4bd39635076df6",
                "responses": {
                    "302": {
                        "description": "Redirect to github.com/login/oauth/authorize"
                    }
                }
            }
        },
        "/api/auth/github/callback": {
            "get": {
                "tags": [
                    "Auth - GitHub"
                ],
                "summary": "GitHub OAuth callback",
                "description": "Creates/updates the user and returns a Sanctum token.",
                "operationId": "aa7596359b6d521c12ac1d071337548f",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful"
                    },
                    "401": {
                        "description": "GitHub authentication failed"
                    }
                }
            }
        },
        "/api/github/repositories": {
            "get": {
                "tags": [
                    "GitHub"
                ],
                "summary": "Get GitHub repositories",
                "description": "Retrieve the authenticated user's GitHub repositories using their connected GitHub account.",
                "operationId": "7b25f5a0349e0e741b37ff19e124d638",
                "responses": {
                    "201": {
                        "description": "Project created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "GitHub account is not connected or validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No GitHub account connected."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List all projects",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "List of projects"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a project",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "My Project",
                                "description": "My project description",
                                "repositories": [
                                    {
                                        "github_repo_id": "123456789",
                                        "name": "my-project",
                                        "owner": "john",
                                        "github_url": "https://github.com/john/my-project",
                                        "github_created_at": "2026-08-20T10:00:00Z",
                                        "github_updated_at": "2026-08-23T15:30:00Z"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully"
                    },
                    "422": {
                        "description": "Validation error or GitHub account not connected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh the access token",
                "description": "Requires a valid, non-expired refresh token sent as Bearer token. Rotates both tokens.",
                "operationId": "e760c9efdbf86b4c43f957e7849a7bc4",
                "responses": {
                    "200": {
                        "description": "New token pair issued"
                    },
                    "401": {
                        "description": "Refresh token missing, invalid or expired"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke current auth tokens",
                "operationId": "aa30070c226ed03cad6e2b80ac458520",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth - GitHub",
            "description": "User authentication via GitHub"
        },
        {
            "name": "Auth",
            "description": "Token lifecycle"
        },
        {
            "name": "GitHub",
            "description": "GitHub"
        },
        {
            "name": "Projects",
            "description": "Projects"
        }
    ]
}